✨ feat #82: useCurrentRegion 커스텀 훅 생성
